Reset dispnet.training_mode to True after testing
dispnet.training_mode determines whether the input dataset
is TRAIN or TEST. feed_dict[dispnet.training_mode] is set to
False for testing with TEST dataset when step % test_step == 0.
However, I find no sentences setting it back to True for training
with TRAIN dataset.
I tested on a simplified model and think this is a BUG.
So reset it to True after each test.